- Added background location and acceleration detection to app - Changed datatypes in entities

+++ Client/app/src/main/java/com/hackathon/optfit/entities/AccelerationReading.java
@@ -5,8 +5,8 @@ public class AccelerationReading {
     public Integer id;
     public String timeStamp;
     public Integer userId;
-    public Integer x;
-    public Integer y;
-    public Integer z;
+    public float x;
+    public float y;
+    public float z;
 
 }
