+++ Server/OpFit.Entities/Readings/Acceleration.cs
@@ -16,8 +16,8 @@ namespace OpFit.Entities.Readings
         public User User { get; set; } 
         #endregion
 
-        public int X { get; set; }
-        public int Y { get; set; }
-        public int Z { get; set; }
+        public float X { get; set; }
+        public float Y { get; set; }
+        public float Z { get; set; }
     }
 }
